// src/doctor/enums/specialization.enum.ts
export enum Specialization {
  CARDIOLOGIST = 'CARDIOLOGIST',
  DERMATOLOGIST = 'DERMATOLOGIST',
  NEUROLOGIST = 'NEUROLOGIST',
  GASTROENTEROLOGIST = 'GASTROENTEROLOGIST',
  ORTHOPEDIC_SURGEON = 'ORTHOPEDIC_SURGEON',
  PEDIATRICIAN = 'PEDIATRICIAN',
  PSYCHIATRIST = 'PSYCHIATRIST',
  ENT_SPECIALIST = 'ENT_SPECIALIST',
  OPHTHALMOLOGIST = 'OPHTHALMOLOGIST',
  GYNECOLOGIST = 'GYNECOLOGIST',
  UROLOGIST = 'UROLOGIST',
  PULMONOLOGIST = 'PULMONOLOGIST',
  GENERAL_PHYSICIAN = 'GENERAL_PHYSICIAN',
  ENDOCRINOLOGIST = 'ENDOCRINOLOGIST',
  ONCOLOGIST = 'ONCOLOGIST',
  NEPHROLOGIST = 'NEPHROLOGIST',
}